+++ metrics/src/io/pedestal/metrics.clj
@@ -1,4 +1,4 @@
-; Copyright 2023 NuBank NA
+; Copyright 2023 Nubank NA
 
 ; The use and distribution terms for this software are covered by the
 ; Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0)